Modificaciones al código en useUsers para saber si el name está ya en el Rank. Y al terminar el test, si ya existe, se hace un overwrite

+++ quiz-app/src/hooks/useUsers.js
@@ -1,11 +1,30 @@
+import { database } from "../firebase.js";
+
 export const useUsers = () => {
   const handleSubmit = (event) => {
     event.preventDefault();
 
     const name = event.target.name.value;
+
+    if (isNameInRank(name) === true) {
+      const confirm = window.confirm(
+        "There is already an user with that name. Do you want to overwrite it?"
+      );
+
+      if (!confirm) {
+        return;
+      }
+    }
+
     sessionStorage.setItem("name", name);
 
     window.location.href = "/quiz";
+  };
+
+  const isNameInRank = async (name) => {
+    const queryRef = database.collection("rank");
+    const querySnapshot = await queryRef.where("nombre", "==", name).get();
+    return !querySnapshot.empty;
   };
 
   return { handleSubmit };
